3.9-3.11 Connect app to backend from part3
Made small edits such as address changes to make sure the app works
with the mock backend from part2 and full backend from part3.

The previous state of part2 was captured to a separate branch to make
sure that no valuable info is lost. Unfortunately the exercise
submission was already made using the master branch as a link.